Fix typo in oauth2 test namespace and use named logger for FoursquareProvider
