+++ test/com/mohiva/play/silhouette/core/providers/oauth2/FoursquareProviderSpec.scala
@@ -13,23 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mohiva.play.silhouette.core.providers.oauh2
+package com.mohiva.play.silhouette.core.providers.oauth2
 
 import test.Helper
 import java.util.UUID
+import play.api.libs.json.Json
 import play.api.libs.ws.{ Response, WS }
 import play.api.test.{ FakeRequest, WithApplication }
 import scala.concurrent.Future
 import com.mohiva.play.silhouette.core.providers._
-import com.mohiva.play.silhouette.core.providers.oauth2.FacebookProvider
 import com.mohiva.play.silhouette.core.{ LoginInfo, AuthenticationException }
-import FacebookProvider._
+import FoursquareProvider._
 import OAuth2Provider._
 
 /**
- * Test case for the [[com.mohiva.play.silhouette.core.providers.oauth2.FacebookProvider]] class.
+ * Test case for the [[com.mohiva.play.silhouette.core.providers.oauth2.FoursquareProvider]] class.
  */
-class FacebookProviderSpec extends OAuth2ProviderSpec {
+class FoursquareProviderSpec extends OAuth2ProviderSpec {
 
   "The authenticate method" should {
     "throw AuthenticationException if OAuth2Info can be build because of an unexpected response" in new WithApplication with Context {
@@ -38,13 +38,13 @@ class FacebookProviderSpec extends OAuth2ProviderSpec {
       val requestHolder = mock[WS.WSRequestHolder]
       val response = mock[Response]
       implicit val req = FakeRequest(GET, "?" + Code + "=my.code&" + State + "=" + state).withSession(CacheKey -> cacheID)
-      response.body returns ""
+      response.json returns Json.obj()
       requestHolder.post[Map[String, Seq[String]]](any)(any, any) returns Future.successful(response)
       cacheLayer.get[String](cacheID) returns Future.successful(Some(state))
       httpLayer.url(oAuthSettings.accessTokenURL) returns requestHolder
 
       await(provider.authenticate()) must throwAn[AuthenticationException].like {
-        case e => e.getMessage must equalTo(InvalidResponseFormat.format(provider.id, ""))
+        case e => e.getMessage must startWith(InvalidResponseFormat.format(provider.id, ""))
       }
     }
 
@@ -54,20 +54,19 @@ class FacebookProviderSpec extends OAuth2ProviderSpec {
       val requestHolder = mock[WS.WSRequestHolder]
       val response = mock[Response]
       implicit val req = FakeRequest(GET, "?" + Code + "=my.code&" + State + "=" + state).withSession(CacheKey -> cacheID)
-      response.body returns AccessToken + "=my.access.token&" + Expires + "=1"
-      response.json returns Helper.loadJson("providers/oauth2/facebook.error.json")
+      response.json returns oAuthInfo thenReturns Helper.loadJson("providers/oauth2/foursquare.error.json")
       requestHolder.post[Map[String, Seq[String]]](any)(any, any) returns Future.successful(response)
       requestHolder.get() returns Future.successful(response)
       cacheLayer.get[String](cacheID) returns Future.successful(Some(state))
       httpLayer.url(oAuthSettings.accessTokenURL) returns requestHolder
-      httpLayer.url(API.format("my.access.token")) returns requestHolder
+      httpLayer.url(API.format("my.access.token", DefaultAPIVersion)) returns requestHolder
 
       await(provider.authenticate()) must throwAn[AuthenticationException].like {
         case e => e.getMessage must equalTo(SpecifiedProfileError.format(
           provider.id,
-          "An active access token must be used to query information about the current user.",
-          "OAuthException",
-          2500))
+          400,
+          Some("param_error"),
+          Some("Must provide a valid user ID or 'self.'")))
       }
     }
 
@@ -77,13 +76,12 @@ class FacebookProviderSpec extends OAuth2ProviderSpec {
       val requestHolder = mock[WS.WSRequestHolder]
       val response = mock[Response]
       implicit val req = FakeRequest(GET, "?" + Code + "=my.code&" + State + "=" + state).withSession(CacheKey -> cacheID)
-      response.body returns AccessToken + "=my.access.token&" + Expires + "=1"
-      response.json throws new RuntimeException("")
+      response.json returns oAuthInfo thenThrows new RuntimeException("")
       requestHolder.post[Map[String, Seq[String]]](any)(any, any) returns Future.successful(response)
       requestHolder.get() returns Future.successful(response)
       cacheLayer.get[String](cacheID) returns Future.successful(Some(state))
       httpLayer.url(oAuthSettings.accessTokenURL) returns requestHolder
-      httpLayer.url(API.format("my.access.token")) returns requestHolder
+      httpLayer.url(API.format("my.access.token", DefaultAPIVersion)) returns requestHolder
 
       await(provider.authenticate()) must throwAn[AuthenticationException].like {
         case e => e.getMessage must equalTo(UnspecifiedProfileError.format(provider.id))
@@ -96,69 +94,114 @@ class FacebookProviderSpec extends OAuth2ProviderSpec {
       val requestHolder = mock[WS.WSRequestHolder]
       val response = mock[Response]
       implicit val req = FakeRequest(GET, "?" + Code + "=my.code&" + State + "=" + state).withSession(CacheKey -> cacheID)
-      response.body returns AccessToken + "=my.access.token&" + Expires + "=1"
-      response.json returns Helper.loadJson("providers/oauth2/facebook.success.json")
+      response.json returns oAuthInfo thenReturns Helper.loadJson("providers/oauth2/foursquare.success.json")
       requestHolder.post[Map[String, Seq[String]]](any)(any, any) returns Future.successful(response)
       requestHolder.get() returns Future.successful(response)
       cacheLayer.get[String](cacheID) returns Future.successful(Some(state))
       httpLayer.url(oAuthSettings.accessTokenURL) returns requestHolder
-      httpLayer.url(API.format("my.access.token")) returns requestHolder
+      httpLayer.url(API.format("my.access.token", DefaultAPIVersion)) returns requestHolder
 
       await(provider.authenticate()) must beRight.like {
         case p =>
           p must be equalTo new SocialProfile(
-            loginInfo = LoginInfo(provider.id, "134405962728980"),
+            loginInfo = LoginInfo(provider.id, "13221052"),
             firstName = Some("Apollonia"),
             lastName = Some("Vanova"),
-            fullName = Some("Apollonia Vanova"),
             email = Some("apollonia.vanova@watchmen.com"),
-            avatarURL = Some("https://fbcdn-sphotos-g-a.akamaihd.net/hphotos-ak-ash2/t1/36245_155530314499277_2350717_n.jpg?lvh=1")
+            avatarURL = Some("https://irs0.4sqi.net/img/user/100x100/blank_girl.png")
           )
       }
     }
 
-    "store the auth info with expires value if the authentication was successful" in new WithApplication with Context {
+    "return the social profile if API is deprecated" in new WithApplication with Context {
       val cacheID = UUID.randomUUID().toString
       val state = UUID.randomUUID().toString
       val requestHolder = mock[WS.WSRequestHolder]
       val response = mock[Response]
       implicit val req = FakeRequest(GET, "?" + Code + "=my.code&" + State + "=" + state).withSession(CacheKey -> cacheID)
-      response.body returns AccessToken + "=my.access.token&" + Expires + "=1"
-      response.json returns Helper.loadJson("providers/oauth2/facebook.success.json")
+      response.json returns oAuthInfo thenReturns Helper.loadJson("providers/oauth2/foursquare.deprecated.json")
       requestHolder.post[Map[String, Seq[String]]](any)(any, any) returns Future.successful(response)
       requestHolder.get() returns Future.successful(response)
       cacheLayer.get[String](cacheID) returns Future.successful(Some(state))
       httpLayer.url(oAuthSettings.accessTokenURL) returns requestHolder
-      httpLayer.url(API.format("my.access.token")) returns requestHolder
+      httpLayer.url(API.format("my.access.token", DefaultAPIVersion)) returns requestHolder
 
-      await(provider.authenticate())
-      there was one(authInfoService).save[OAuth2Info](LoginInfo(provider.id, "134405962728980"), OAuth2Info(
-        accessToken = "my.access.token",
-        tokenType = None,
-        expiresIn = Some(1),
-        refreshToken = None))
+      await(provider.authenticate()) must beRight.like {
+        case p =>
+          p must be equalTo new SocialProfile(
+            loginInfo = LoginInfo(provider.id, "13221052"),
+            firstName = Some("Apollonia"),
+            lastName = Some("Vanova"),
+            email = Some("apollonia.vanova@watchmen.com"),
+            avatarURL = Some("https://irs0.4sqi.net/img/user/100x100/blank_girl.png")
+          )
+      }
     }
 
-    "store the auth info without expires value if the authentication was successful" in new WithApplication with Context {
+    "handle the custom API version property" in new WithApplication with Context {
+      val customProperties = Map(APIVersion -> "20120101")
       val cacheID = UUID.randomUUID().toString
       val state = UUID.randomUUID().toString
       val requestHolder = mock[WS.WSRequestHolder]
       val response = mock[Response]
       implicit val req = FakeRequest(GET, "?" + Code + "=my.code&" + State + "=" + state).withSession(CacheKey -> cacheID)
-      response.body returns AccessToken + "=my.access.token"
-      response.json returns Helper.loadJson("providers/oauth2/facebook.success.json")
+      override lazy val provider = new FoursquareProvider(
+        authInfoService,
+        cacheLayer,
+        httpLayer,
+        oAuthSettings.copy(customProperties = customProperties)
+      )
+
+      response.json returns oAuthInfo thenReturns Helper.loadJson("providers/oauth2/foursquare.success.json")
       requestHolder.post[Map[String, Seq[String]]](any)(any, any) returns Future.successful(response)
       requestHolder.get() returns Future.successful(response)
       cacheLayer.get[String](cacheID) returns Future.successful(Some(state))
       httpLayer.url(oAuthSettings.accessTokenURL) returns requestHolder
-      httpLayer.url(API.format("my.access.token")) returns requestHolder
+      httpLayer.url(API.format("my.access.token", "20120101")) returns requestHolder
 
-      await(provider.authenticate())
-      there was one(authInfoService).save[OAuth2Info](LoginInfo(provider.id, "134405962728980"), OAuth2Info(
-        accessToken = "my.access.token",
-        tokenType = None,
-        expiresIn = None,
-        refreshToken = None))
+      await(provider.authenticate()) must beRight.like {
+        case p =>
+          p must be equalTo new SocialProfile(
+            loginInfo = LoginInfo(provider.id, "13221052"),
+            firstName = Some("Apollonia"),
+            lastName = Some("Vanova"),
+            email = Some("apollonia.vanova@watchmen.com"),
+            avatarURL = Some("https://irs0.4sqi.net/img/user/100x100/blank_girl.png")
+          )
+      }
+    }
+
+    "handle the custom avatar resolution property" in new WithApplication with Context {
+      val customProperties = Map(AvatarResolution -> "150x150")
+      val cacheID = UUID.randomUUID().toString
+      val state = UUID.randomUUID().toString
+      val requestHolder = mock[WS.WSRequestHolder]
+      val response = mock[Response]
+      implicit val req = FakeRequest(GET, "?" + Code + "=my.code&" + State + "=" + state).withSession(CacheKey -> cacheID)
+      override lazy val provider = new FoursquareProvider(
+        authInfoService,
+        cacheLayer,
+        httpLayer,
+        oAuthSettings.copy(customProperties = customProperties)
+      )
+
+      response.json returns oAuthInfo thenReturns Helper.loadJson("providers/oauth2/foursquare.success.json")
+      requestHolder.post[Map[String, Seq[String]]](any)(any, any) returns Future.successful(response)
+      requestHolder.get() returns Future.successful(response)
+      cacheLayer.get[String](cacheID) returns Future.successful(Some(state))
+      httpLayer.url(oAuthSettings.accessTokenURL) returns requestHolder
+      httpLayer.url(API.format("my.access.token", DefaultAPIVersion)) returns requestHolder
+
+      await(provider.authenticate()) must beRight.like {
+        case p =>
+          p must be equalTo new SocialProfile(
+            loginInfo = LoginInfo(provider.id, "13221052"),
+            firstName = Some("Apollonia"),
+            lastName = Some("Vanova"),
+            email = Some("apollonia.vanova@watchmen.com"),
+            avatarURL = Some("https://irs0.4sqi.net/img/user/150x150/blank_girl.png")
+          )
+      }
     }
   }
 
@@ -178,16 +221,15 @@ class FacebookProviderSpec extends OAuth2ProviderSpec {
      * The OAuth2 settings.
      */
     lazy val oAuthSettings = OAuth2Settings(
-      authorizationURL = "https://graph.facebook.com/oauth/authorize",
-      accessTokenURL = "https://graph.facebook.com/oauth/access_token",
+      authorizationURL = "https://foursquare.com/oauth2/authenticate",
+      accessTokenURL = "https://foursquare.com/oauth2/access_token",
       redirectURL = "https://www.mohiva.com",
       clientID = "my.client.id",
-      clientSecret = "my.client.secret",
-      scope = Some("email"))
+      clientSecret = "my.client.secret")
 
     /**
      * The provider to test.
      */
-    lazy val provider = new FacebookProvider(authInfoService, cacheLayer, httpLayer, oAuthSettings)
+    lazy val provider = new FoursquareProvider(authInfoService, cacheLayer, httpLayer, oAuthSettings)
   }
 }
